% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getgac.R
\name{getgac}
\alias{getgac}
\title{A web scraping function}
\usage{
getgac(year, month)
}
\description{
This function allows you to scrape data from Global Affairs Canada softwood lumber Monthly Export Reports (Canada-US)
}
\examples{
jan_2018 <- getgac("2018","01")
}
\keyword{exports}
